set loading time to 15s
